Extract paramiko SSH server test helper into testing.py
Moves _PasswordAuthServer and _run_ssh_server from test_password_auth.py
into testing.py (standard non-fixture test utility module per repo
convention). Test file is now focused on assertions; server plumbing
lives with other test helpers and is reusable.
